+++ pyxel/configuration/configuration.py
@@ -11,7 +11,7 @@ from collections.abc import Iterator, Mapping, Sequence
 from dataclasses import dataclass
 from pathlib import Path
 from shutil import copy2
-from typing import IO, TYPE_CHECKING, Any, Optional, Union
+from typing import IO, TYPE_CHECKING, Any, Literal, Optional, Union
 
 from pyxel import __version__ as version
 from pyxel.detectors import (
@@ -160,6 +160,102 @@ def load_yaml(stream: str | IO) -> Any:
 
     result = yaml.load(stream, Loader=yaml.SafeLoader)
     return result
+
+
+def build_configuration(
+    detector_type: Literal["CCD", "CMOS", "MKID", "APD"],
+    num_rows: int,
+    num_cols: int,
+) -> Configuration:
+    """Build a default Pyxel ``Configuration`` object for a specified detector type.
+
+    Parameters
+    ----------
+    detector_type : 'CCD', 'CMOS', 'MKID', 'APD'
+        Type of detector for which the configuration should be built.
+    num_rows : int
+        Number of pixel row in the detector.
+    num_cols : int
+        Number of pixel columns in the detector.
+
+    Returns
+    -------
+    Configuration
+        A fully defined ``Configuration`` object with pre-filled basic models.
+
+    Examples
+    --------
+    >>> import pyxel
+    >>> config = pyxel.build_configuration(
+    ...     detector_type="CCD",
+    ...     num_rows=512,
+    ...     num_cols=512,
+    ... )
+    >>> config.detector.geometry.row
+    512
+    >>> result = pyxel.run_mode(config)
+    """
+    match detector_type:
+        case "CCD":
+            # Define a default configuration for a CCD detector
+            config = Configuration(
+                exposure=Exposure(readout=Readout()),
+                ccd_detector=CCD(
+                    geometry=CCDGeometry(row=num_rows, col=num_cols),
+                    environment=Environment(),
+                    characteristics=Characteristics(
+                        quantum_efficiency=0.8,
+                        charge_to_volt_conversion=1e-6,
+                        pre_amplification=100.0,
+                        adc_bit_resolution=16,
+                        adc_voltage_range=(0.0, 10.0),
+                    ),
+                ),
+                pipeline=DetectionPipeline(
+                    # Generate photons with a USAF pattern
+                    photon_collection=[
+                        ModelFunction(
+                            name="usaf_illumination",
+                            func="pyxel.models.photon_collection.usaf_illumination",
+                        )
+                    ],
+                    # Convert photons to electrons
+                    charge_generation=[
+                        ModelFunction(
+                            name="simple_conversion",
+                            func="pyxel.models.charge_generation.simple_conversion",
+                        )
+                    ],
+                    charge_collection=[
+                        ModelFunction(
+                            name="simple_collection",
+                            func="pyxel.models.charge_collection.simple_collection",
+                        )
+                    ],
+                    # Convert electrons to volt
+                    charge_measurement=[
+                        ModelFunction(
+                            name="simple_measurement",
+                            func="pyxel.models.charge_measurement.simple_measurement",
+                        )
+                    ],
+                    readout_electronics=[
+                        ModelFunction(
+                            name="simple_amplifier",
+                            func="pyxel.models.readout_electronics.simple_amplifier",
+                        ),
+                        ModelFunction(
+                            name="simple_adc",
+                            func="pyxel.models.readout_electronics.simple_adc",
+                        ),
+                    ],
+                ),
+            )
+
+        case _:
+            raise NotImplementedError("Currently only 'CCD' detector is implemented.")
+
+    return config
 
 
 def to_exposure_outputs(dct: dict | None) -> ExposureOutputs: